fix(native): close Windows process boundaries safely
