+++ ci/test-suite-ratchet-baseline.json
@@ -1,5 +1,5 @@
 {
   "schema_version": 1,
-  "maximum_collected_cases": 16398,
-  "maximum_test_source_lines": 354398
+  "maximum_collected_cases": 16407,
+  "maximum_test_source_lines": 354734
 }
